BUG: Assign loaded masks to correct variable
Loaded masks were incorrectly stored in the `image_nodes` variable, as opposed to the intended `mask_nodes` variable, causing slicer to not recognize loaded masks as such when saving the case.

Additionally, find other masks by looking for the label map nodes by class, as opposed to the pattern '-label', and store them in the same directory as the loaded image (instead of `root`, which may be just the one folder, thereby potentially resulting in file-overwriting.