ENH: Observe MRMLScene EndCloseEvent
Observe this event to set the current case to None. This ensures that SlicerBatch does not attempt to save and close the case if the user has already closed the case manually (e.g. by CTRL+W).

Additionally, change the order of the function to better group them together and add more extensive documentation to the function where necessary.