ENH: Change input file selector to ctkPathLineEdit
Change the input from a simple text box to a path line edit, which adds autocomplete and a browse button to facilitate selection of the input file.

Additionally, add a checkbox to control the auto switching to the editor module (default: switching_on)

When additional images per case are specified, load those in all three viewers, instead of just the red slice viewer.