Switch to loading cases from the table node
Update interface to load the batch from the selected table node instead of loading and storing the csv input file in a variable. This way the user can preview the batch and optionally edit the table prior to loading.

Interesting note: Because part of closing the case is instructing slicer to close the scene, the table is lost when user switches to a different case. This also discards any changes the users makes in the table.

The input file selector retains its value, enabling the user to easily reload the input from file.