ENH: Prevent discarding of table during iteration
As part of the switch to a different case, the slicer scene is closed. However, this scene includes the input table.
Upon each close, if the tableNode is not None, readd the table.

Additionally, fix a small bug when loading a mask. This was due to refactoring from a previous version (which sent a dictionary representing the case and the corresponding column names to the logic.)