STYL: Update documentation and console output
Update documentation to include description of the expected python console output.
Add parsing of columns `patient` or `ID` (which is treated as a case identifier and only used to report which case is being loaded).

Fix some small bugs in formatting the output.

Update the docstrings for `loadCase` to better reflect what the function does (it was incorrect, as the reflected code function from before refactoring to use table node as input).