ENH: Use GetStorageNode().GetFileName() to get location of loaded table file.
Previously, the directory containing the table file was determined using the current path in the `input CSV path` selector. However, this did not work when a table was loaded differently into the slicer scene (e.g. using a drag an drop operation).
Upon starting the batch, check if the current selected table has a storage node which contains a filename. If so, use the directory containing the file as the `csv_dir`, which is used as a reference when parsing relative paths.