ENH: Use Segment Editor instead of Editor in Case Iterator
The Editor module is considered legacy. It is already in the legacy category and does not appear in the main module list in 4.10 and will be removed from 5.0. Also, Segment Editor is considered to be a significantly more advanced version of Editor, so it makes sense to use it when doing segmentations with Slicer.

Labelmaps can still be specified as input mask. In that case, it will be imported into a segmentation node and the labelmap node removed.