BUG: Fix a memory leak when loading label volumes
When a mask is detected to be a label volume, it is loaded and converted to segmentation. Because the segmentation is a new node, a storage node is also added (which is used to determine the storage location upon saving the segmentation).
However, when creating a storage node using `CreateDefaultStorageNode()`, the reference count is increased. To allow correct GC, UnRegister the node after it is linked to the segmentation node.

Additionally, fix a bug due to the renaming of the IteratorLogicBase.