Add GUI and logic to use DICOM input
Update the widget to allow the user to choose what type of input to use.
Add a new iterator widget and logic class to work with slicer's DICOM browser. In this new input type, masks are still loaded and saved to local files, but images are retrieved via SeriesUID specified in a table.