ENH: Pass reader to iterator logic
Value for reader is only used by iterator logic to save the nodes. Pass it to the iterator logic class via function `startBatch` to make it available at load.
That way it can also be used for additional input validation (e.g. checking whether part of the batch is already done for this reader.