BUG: Fix Access Violation error on Slicer close
When references to widgets are retained when Slicer is closed, the widgets cannot be properly destroyed.
In this case, Qt destroys the widgets, including references. Then, when the widget class is destroyed after, the variables referencing the destroyed widgets point to invalid memory addresses and the error occurs.

To prevent this error, ensure the python classes holding the widget references are destroyed before Qt cleans up the parent widget. This is achieved by removing the references to the iterator widgets in the cleanup function.